finish explore page and finsh createModal!

+++ server_ichgram/controllers/likeComtroller.js
@@ -1,5 +1,34 @@
-import Like from "../models/Like";
+import Like from "../models/Like.js";
 import Post from "../models/Post.js";
+
+export const getLikes = async (req, res) => {
+  try {
+    const { postId } = req.query;
+
+    if (!postId) {
+      return res.status(400).json({
+        message: "Post ID is required",
+        success: false,
+      });
+    }
+
+    const likes = await Like.find({ postId }).populate(
+      "userId",
+      "username avatar"
+    );
+
+    res.status(200).json({
+      success: true,
+      likes,
+    });
+  } catch (error) {
+    res.status(500).json({
+      message: "Server error",
+      success: false,
+      error: error.message,
+    });
+  }
+};
 
 export const addLike = async (req, res) => {
   try {
@@ -28,7 +57,7 @@ export const addLike = async (req, res) => {
         success: false,
       });
     }
-    await Like.create({
+    const like = await Like.create({
       postId,
       userId,
     });
@@ -38,6 +67,7 @@ export const addLike = async (req, res) => {
     res.status(200).json({
       message: "Like added",
       success: true,
+      like,
     });
   } catch (error) {
     res.status(500).json({
@@ -58,6 +88,15 @@ export const removeLike = async (req, res) => {
         success: false,
       });
     }
+    const post = await Post.findById(postId);
+
+    if (!post) {
+      return res.status(404).json({
+        message: "Post not found",
+        success: false,
+      });
+    }
+
     const like = await Like.findOneAndDelete({
       postId,
       userId: req.user._id,
@@ -68,11 +107,13 @@ export const removeLike = async (req, res) => {
         success: false,
       });
     }
+
     post.likesCount -= 1;
     await post.save();
     res.status(200).json({
       message: "Like deleted",
       success: true,
+      like,
     });
   } catch (error) {
     res.status(500).json({
